feat: handleError 함수 구현
